Remove all Compass/Celery/Sync code
