+++ .env-template
@@ -86,13 +86,6 @@ OIDC_WELL_KNOWN_ENDPOINT=<OIDC_WELL_KNOWN_ENDPOINT>
 SCIM_USER=<SCIM_USER>
 SCIM_PASSWORD=<SCIM_PASSWORD>
 
-# Compass
-
-COHERE_COMPASS_USERNAME=
-COHERE_COMPASS_PASSWORD=
-COHERE_COMPASS_API_URL=
-COHERE_COMPASS_PARSER_URL=
-
 # Google Drive
 
 GOOGLE_DRIVE_CLIENT_ID=
