+++ src/backend/alembic/versions/2024_09_18_803535b4e118_.py
@@ -1,0 +1,34 @@
+"""
+
+Revision ID: 803535b4e118
+Revises: ac3933258035
+Create Date: 2024-09-18 17:04:40.969832
+
+"""
+from typing import Sequence, Union
+
+import sqlalchemy as sa
+from alembic import op
+from sqlalchemy.dialects import postgresql
+
+# revision identifiers, used by Alembic.
+revision: str = '803535b4e118'
+down_revision: Union[str, None] = 'ac3933258035'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.alter_column('agent_tool_metadata', 'artifacts',
+               existing_type=postgresql.JSON(astext_type=sa.Text()),
+               nullable=False)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.alter_column('agent_tool_metadata', 'artifacts',
+               existing_type=postgresql.JSON(astext_type=sa.Text()),
+               nullable=True)
+    # ### end Alembic commands ###
