+++ src/backend/config/secrets.template.yaml
@@ -20,9 +20,6 @@ tools:
     api_key:
   wolfram_alpha:
     app_id:
-  compass:
-    username:
-    password:
   google_drive:
     client_id:
     client_secret:
@@ -39,8 +36,3 @@ auth:
     client_id:
     client_secret:
     well_known_endpoint:
-compass:
-  username:
-  password:
-  api_url:
-  parser_url:
