+++ src/backend/database_models/__init__.py
@@ -1,6 +1,5 @@
 # ruff: noqa
 from backend.database_models.agent import *
-from backend.database_models.agent_task import *
 from backend.database_models.agent_tool_metadata import *
 from backend.database_models.base import *
 from backend.database_models.blacklist import *
