+++ src/backend/services/agent.py
@@ -1,13 +1,10 @@
-import pickle
 
 from fastapi import HTTPException
 
 from backend.crud import agent as agent_crud
 from backend.crud import agent_tool_metadata as agent_tool_metadata_crud
 from backend.database_models.agent import Agent, AgentToolMetadata
-from backend.database_models.agent_task import SyncCeleryTaskMeta
 from backend.database_models.database import DBSessionDep
-from backend.schemas.agent import AgentTaskResponse
 
 TASK_TRACE_PREVIEW_LIMIT = 200
 
@@ -47,25 +44,3 @@ def raise_db_error(e: Exception, type: str, name: str):
         )
 
     raise HTTPException(status_code=500, detail=str(e))
-
-
-def parse_task(t: SyncCeleryTaskMeta) -> AgentTaskResponse:
-    result = None
-    exception_snippet = None
-    if t.status == "SUCCESS":
-        result = pickle.loads(t.result)
-    if t.status == "FAILURE":
-        trace_lines = t.traceback.split("\n")
-        if len(trace_lines) >= 2:
-            # first 200 characters of the exception
-            exception_snippet = trace_lines[-2][:TASK_TRACE_PREVIEW_LIMIT] + "...check logs for details"
-
-    return AgentTaskResponse(
-        task_id=t.task_id,
-        status=t.status,
-        name=t.name,
-        retries=t.retries,
-        result=result,
-        exception_snippet=exception_snippet,
-        date_done=str(t.date_done),
-    )
