--- src/backend/services/sync/config.py
@@ -1,8 +1,0 @@
-import httpx
-from dotenv import load_dotenv
-
-load_dotenv()
-
-
-class Configuration:
-    HTTPX_CLIENT: httpx.Client = httpx.Client(http2=True, timeout=300)
