+++ src/backend/tests/unit/configuration.yaml
@@ -19,13 +19,10 @@ tools:
   enabled_tools:
   python_interpreter:
     url:
-  compass:
-    api_url:
-    parser_url:
 feature_flags:
   # Experimental features
   use_experimental_langchain: false
-  use_agents_view: true
+  use_agents_view: false
   # Community features
   use_community_features: true
 auth:
