+++ src/backend/tests/unit/routers/test_conversation.py
@@ -495,7 +495,7 @@ def test_search_conversations_missing_user_id(
 
 # FILES
 def test_list_files(
-    session_client: TestClient, session: Session, user: User, mock_compass_settings
+    session_client: TestClient, session: Session, user: User
 ) -> None:
     conversation = get_factory("Conversation", session).create(user_id=user.id)
     files = [
@@ -531,7 +531,7 @@ def test_list_files(
 
 
 def test_list_files_no_files(
-    session_client: TestClient, session: Session, user: User, mock_compass_settings
+    session_client: TestClient, session: Session, user: User
 ) -> None:
     conversation = get_factory("Conversation", session).create(user_id=user.id)
     response = session_client.get(
@@ -544,7 +544,7 @@ def test_list_files_no_files(
 
 
 def test_list_files_missing_user_id(
-    session_client: TestClient, session: Session, user: User, mock_compass_settings
+    session_client: TestClient, session: Session, user: User
 ) -> None:
     conversation = get_factory("Conversation", session).create(user_id=user.id)
     response = session_client.get(f"/v1/conversations/{conversation.id}/files")
@@ -554,7 +554,7 @@ def test_list_files_missing_user_id(
 
 
 def test_batch_upload_file_existing_conversation(
-    session_client: TestClient, session: Session, user, mock_compass_settings
+    session_client: TestClient, session: Session, user
 ) -> None:
     file_paths = {
         "Mariana_Trench.pdf": "src/backend/tests/unit/test_data/Mariana_Trench.pdf",
@@ -595,7 +595,7 @@ def test_batch_upload_file_existing_conversation(
 
 
 def test_batch_upload_file_nonexistent_conversation_creates_new_conversation(
-    session_client: TestClient, session: Session, user, mock_compass_settings
+    session_client: TestClient, session: Session, user
 ) -> None:
     file_paths = {
         "Mariana_Trench.pdf": "src/backend/tests/unit/test_data/Mariana_Trench.pdf",
@@ -649,7 +649,7 @@ def test_batch_upload_file_nonexistent_conversation_creates_new_conversation(
 
 
 def test_batch_upload_file_nonexistent_conversation_fails_if_user_id_not_provided(
-    session_client: TestClient, session: Session, user: User, mock_compass_settings
+    session_client: TestClient, session: Session, user: User
 ) -> None:
     file_paths = {
         "Mariana_Trench.pdf": "src/backend/tests/unit/test_data/Mariana_Trench.pdf",
@@ -672,7 +672,6 @@ def test_delete_file(
     session_client: TestClient,
     session: Session,
     user: User,
-    mock_compass_settings,
 ) -> None:
     conversation = get_factory("Conversation", session).create(user_id=user.id)
     files = [
@@ -722,7 +721,7 @@ def test_delete_file(
 
 
 def test_fail_delete_nonexistent_file(
-    session_client: TestClient, session: Session, user: User, mock_compass_settings
+    session_client: TestClient, session: Session, user: User
 ) -> None:
     conversation = get_factory("Conversation", session).create(user_id=user.id)
     response = session_client.delete(
@@ -735,7 +734,7 @@ def test_fail_delete_nonexistent_file(
 
 
 def test_fail_delete_file_missing_user_id(
-    session_client: TestClient, session: Session, user: User, mock_compass_settings
+    session_client: TestClient, session: Session, user: User
 ) -> None:
     conversation = get_factory("Conversation", session).create(user_id=user.id)
     file = get_factory("File", session).create(
