+++ src/backend/tests/unit/secrets.yaml
@@ -20,9 +20,6 @@ tools:
     api_key:
   wolfram_alpha:
     app_id:
-  compass:
-    username:
-    password:
   google_drive:
     client_id:
     client_secret:
