+++ src/backend/tools/files.py
@@ -2,89 +2,15 @@ from enum import StrEnum
 from typing import Any, Dict, List
 
 import backend.crud.file as file_crud
-from backend.compass_sdk import SearchFilter
-from backend.config.settings import Settings
-from backend.services.compass import Compass
-from backend.services.file import get_compass
 from backend.tools.base import BaseTool
 
 
 class FileToolsArtifactTypes(StrEnum):
     local_file = "file"
 
-
-def compass_file_search(
-    file_ids: List[str],
-    conversation_id: str,
-    agent_id: str,
-    query: str,
-    search_limit: int = 5,
-) -> List[Dict[str, Any]]:
-    results = []
-
-    # Note: Compass search currently has an issue where the type of the context is not directly referenced
-    # Temporarily add `.keyword` to workaround this issue.
-    search_filters = [
-        SearchFilter(
-            field="content.file_id.keyword",
-            type=SearchFilter.FilterType.EQ,
-            value=file_id,
-        )
-        for file_id in file_ids
-    ]
-
-    compass = get_compass()
-
-    # Search conversation ID index
-    search_results = compass.invoke(
-            action=Compass.ValidActions.SEARCH,
-            parameters={
-                "index": conversation_id,
-                "query": query,
-                "top_k": search_limit,
-                "filters": search_filters,
-        },
-    )
-
-    if search_results.result:
-        results.extend(search_results.result["hits"])
-
-    # Search agent ID index
-    if agent_id:
-        search_results = compass.invoke(
-            action=Compass.ValidActions.SEARCH,
-            parameters={
-                "index": agent_id,
-                "query": query,
-                "top_k": search_limit,
-                "filters": search_filters,
-            },
-        )
-
-    if search_results.result:
-        results.extend(search_results.result["hits"])
-
-    chunks = sorted(
-        [
-            {
-                "text": chunk["content"]["text"],
-                "score": chunk["score"],
-                "url": result["content"].get("file_name", ""),
-                "title": result["content"].get("file_name", ""),
-            }
-            for result in results
-            for chunk in result["chunks"]
-        ],
-        key=lambda x: x["score"],
-        reverse=True,
-    )[:search_limit]
-
-    return chunks
-
-
 class ReadFileTool(BaseTool):
     """
-    This class reads a file from the file system.
+    Tool to read a file from the file system.
     """
 
     NAME = "read_document"
@@ -103,37 +29,25 @@ class ReadFileTool(BaseTool):
 
         session = kwargs.get("session")
         user_id = kwargs.get("user_id")
-        agent_id = kwargs.get("agent_id")
-        conversation_id = kwargs.get("conversation_id")
         if not file:
             return []
 
         _, file_id = file
-        if Settings().feature_flags.use_compass_file_storage:
-            return compass_file_search(
-                [file_id],
-                conversation_id,
-                agent_id,
-                "*",
-                search_limit=self.SEARCH_LIMIT,
-            )
-        else:
-            retrieved_file = file_crud.get_file(session, file_id, user_id)
-            if not retrieved_file:
-                return []
+        retrieved_file = file_crud.get_file(session, file_id, user_id)
+        if not retrieved_file:
+            return []
 
-            return [
-                {
-                    "text": retrieved_file.file_content,
-                    "title": retrieved_file.file_name,
-                    "url": retrieved_file.file_name,
-                }
-            ]
-
+        return [
+            {
+                "text": retrieved_file.file_content,
+                "title": retrieved_file.file_name,
+                "url": retrieved_file.file_name,
+            }
+        ]
 
 class SearchFileTool(BaseTool):
     """
-    This class searches for a query in a file.
+    Tool to query a list of files.
     """
 
     NAME = "search_file"
@@ -153,8 +67,6 @@ class SearchFileTool(BaseTool):
         query = parameters.get("search_query")
         files = parameters.get("files")
 
-        agent_id = kwargs.get("agent_id")
-        conversation_id = kwargs.get("conversation_id")
         session = kwargs.get("session")
         user_id = kwargs.get("user_id")
 
@@ -162,26 +74,17 @@ class SearchFileTool(BaseTool):
             return []
 
         file_ids = [file_id for _, file_id in files]
-        if Settings().feature_flags.use_compass_file_storage:
-            return compass_file_search(
-                file_ids,
-                conversation_id,
-                agent_id,
-                query,
-                search_limit=self.SEARCH_LIMIT,
-            )
-        else:
-            retrieved_files = file_crud.get_files_by_ids(session, file_ids, user_id)
-            if not retrieved_files:
-                return []
+        retrieved_files = file_crud.get_files_by_ids(session, file_ids, user_id)
+        if not retrieved_files:
+            return []
 
-            results = []
-            for file in retrieved_files:
-                results.append(
-                    {
-                        "text": file.file_content,
-                        "title": file.file_name,
-                        "url": file.file_name,
-                    }
-                )
-            return results
+        results = []
+        for file in retrieved_files:
+            results.append(
+                {
+                    "text": file.file_content,
+                    "title": file.file_name,
+                    "url": file.file_name,
+                }
+            )
+        return results
