+++ src/backend/tools/google_drive/__init__.py
@@ -1,39 +1,11 @@
 from backend.tools.google_drive.auth import GoogleDriveAuth
 from backend.tools.google_drive.constants import (
     GOOGLE_DRIVE_TOOL_ID,
-    GoogleDriveActions,
-)
-from backend.tools.google_drive.sync import (
-    handle_google_drive_sync,
-    list_google_drive_artifacts_file_ids,
-)
-from backend.tools.google_drive.sync.actions import (
-    create,
-    delete,
-    edit,
-    move,
-    permission_change,
-    rename,
-)
-from backend.tools.google_drive.sync.activity import (
-    handle_google_drive_activity_event,
-    query_google_drive_activity,
 )
 from backend.tools.google_drive.tool import GoogleDrive
 
 __all__ = [
     "GoogleDriveAuth",
     "GoogleDrive",
-    "handle_google_drive_activity_event",
     "GOOGLE_DRIVE_TOOL_ID",
-    "GoogleDriveActions",
-    "query_google_drive_activity",
-    "list_google_drive_artifacts_file_ids",
-    "handle_google_drive_sync",
-    "create",
-    "delete",
-    "edit",
-    "move",
-    "permission_change",
-    "rename",
 ]
