--- src/backend/tools/google_drive/sync/actions/__init__.py
@@ -1,15 +1,0 @@
-from backend.tools.google_drive.sync.actions.create import create
-from backend.tools.google_drive.sync.actions.delete import delete
-from backend.tools.google_drive.sync.actions.edit import edit
-from backend.tools.google_drive.sync.actions.move import move
-from backend.tools.google_drive.sync.actions.permission_change import permission_change
-from backend.tools.google_drive.sync.actions.rename import rename
-
-__all__ = [
-    "create",
-    "delete",
-    "edit",
-    "move",
-    "permission_change",
-    "rename",
-]
